Rename scale container
