+++ mobile/asa-go/src/components/ScaleContainer.tsx
